# docs/swagger.yaml
definitions:
  api_types.CalculationDetailedResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/api_types.UserResponse'
      id:
        type: integer
      items:
        description: Главное отличие - наличие Items
        items:
          $ref: '#/definitions/api_types.CalculationItemResponse'
        type: array
      moderator:
        $ref: '#/definitions/api_types.UserResponse'
      status:
        type: string
      submitted_at:
        type: string
      total_profit:
        type: number
    type: object
  api_types.CalculationItemResponse:
    properties:
      coefficient:
        type: number
      final_dividend:
        description: omitempty скроет поле, если оно null
        type: number
      fine:
        type: number
      shareholder:
        $ref: '#/definitions/api_types.ShareholderResponse'
    type: object
  api_types.CalculationItemUpdateRequest:
    properties:
      coefficient:
        type: number
      fine:
        type: number
    type: object
  api_types.CalculationResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/api_types.UserResponse'
      id:
        type: integer
      moderator:
        $ref: '#/definitions/api_types.UserResponse'
      status:
        type: string
      submitted_at:
        type: string
      total_profit:
        type: number
    type: object
  api_types.CalculationUpdateRequest:
    properties:
      total_profit:
        type: number
    type: object
  api_types.CartInfoResponse:
    properties:
      count:
        type: integer
      draft_id:
        type: integer
    type: object
  api_types.ModerationRequest:
    properties:
      status:
        description: '"completed" или "rejected"'
        type: string
    type: object
  api_types.ShareholderRequest:
    properties:
      description:
        type: string
      name:
        type: string
      share:
        type: number
    type: object
  api_types.ShareholderResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        description: Используем *string для nullable
        type: string
      name:
        type: string
      share:
        type: number
    type: object
  api_types.UserLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  api_types.UserRegisterRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  api_types.UserResponse:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
  api_types.UserUpdateRequest:
    properties:
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /dividend-calculations:
    get:
      description: Get all calculations (except draft/deleted). Users see their own,
        moderators see all.
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Start date for filtering (YYYY-MM-DD)
        in: query
        name: from_date
        type: string
      - description: End date for filtering (YYYY-MM-DD)
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_types.CalculationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get list of calculations
      tags:
      - calculations
  /dividend-calculations/{id}:
    delete:
      description: Logically delete a draft calculation.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a draft calculation
      tags:
      - calculations
    get:
      description: Get details of a single calculation by its ID.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationDetailedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a calculation by ID
      tags:
      - calculations
    put:
      consumes:
      - application/json
      description: Update fields of a draft calculation (e.g., total_profit).
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.CalculationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a draft calculation
      tags:
      - calculations
  /dividend-calculations/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Complete or reject a 'submitted' calculation. Requires moderator
        rights.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Moderation action
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.ModerationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationDetailedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator or not a submitted calculation)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Moderate a calculation (Moderator only)
      tags:
      - calculations
  /dividend-calculations/{id}/shareholders/{shareholder_id}:
    delete:
      description: Removes a shareholder from the current user's draft calculation.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder ID
        in: path
        name: shareholder_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a shareholder from a draft calculation
      tags:
      - M-M
    put:
      consumes:
      - application/json
      description: Changes coefficient and fine for a shareholder within the current
        user's draft.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder ID
        in: path
        name: shareholder_id
        required: true
        type: integer
      - description: Parameters to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.CalculationItemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationItemResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a shareholder's parameters in a draft calculation
      tags:
      - M-M
  /dividend-calculations/{id}/submit:
    put:
      description: Change the status of a draft calculation to 'submitted'.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a draft calculation
      tags:
      - calculations
  /dividend-calculations/cart:
    get:
      description: Get current draft ID and item count for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CartInfoResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cart info
      tags:
      - calculations
  /shareholders:
    get:
      description: Get all shareholders with optional name filter. Publicly accessible.
      parameters:
      - description: Filter by shareholder name (case-insensitive)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_types.ShareholderResponse'
            type: array
      summary: Get list of shareholders
      tags:
      - shareholders
    post:
      consumes:
      - application/json
      description: Add a new shareholder to the database. Requires moderator rights.
      parameters:
      - description: Shareholder object
        in: body
        name: shareholder
        required: true
        schema:
          $ref: '#/definitions/api_types.ShareholderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new shareholder (Moderator only)
      tags:
      - shareholders
  /shareholders/{id}:
    delete:
      description: Logically delete a shareholder by its ID. Requires moderator rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a shareholder (Moderator only)
      tags:
      - shareholders
    get:
      description: Get details of a single shareholder by its ID. Publicly accessible.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Shareholder not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a shareholder by ID
      tags:
      - shareholders
    put:
      consumes:
      - application/json
      description: Update an existing shareholder's data. Requires moderator rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder object
        in: body
        name: shareholder
        required: true
        schema:
          $ref: '#/definitions/api_types.ShareholderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a shareholder (Moderator only)
      tags:
      - shareholders
  /shareholders/{id}/add-to-draft:
    post:
      description: Adds a shareholder to the current user's draft calculation.
      parameters:
      - description: Shareholder ID to add
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a shareholder to the draft calculation
      tags:
      - shareholders
  /shareholders/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image and associate it with a shareholder. Requires moderator
        rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID or file
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload an image for a shareholder (Moderator only)
      tags:
      - shareholders
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token.
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api_types.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - users
  /users/logout:
    post:
      description: Invalidates the current user's JWT token by adding it to a blacklist.
      produces:
      - application/json
      responses:
        "200":
          description: Logout status
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log out a user
      tags:
      - users
  /users/me:
    get:
      description: Retrieves the profile information for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/api_types.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates the profile information for the authenticated user.
      parameters:
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api_types.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/api_types.UserResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user's profile
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: User Registration Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api_types.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/api_types.UserResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
swagger: "2.0"